Module leb128::read

A module for reading signed and unsigned integers that have been LEB128 encoded.

Enums

Error
An enumeration of the possible errors that can occur when reading a number encoded with LEB128.

Functions

signed
Read a signed LEB128 number from the given std::io::Readable and return it or an error if reading failed.

unsigned
Read an unsigned LEB128 number from the given std::io::Readable and return it or an error if reading failed.